requirements.txt
scikit-learn
nltk
matplotlib
seaborn
kafka-python
dash
plotly
transformers
tensorflow
numpy
pandas
torch
datasets
torchinfo
tensorboard
tweepy